backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-spectacular
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
drf-spectacular